requirement.txt
opencv-python
dlib
torch
torchvision
transformers
huggingface_hub
numpy
matplotlib
streamlit
fer
pandas
pillow
moviepy
